+++ src/steps/index.ts
@@ -1,8 +1,10 @@
 import { accountSteps } from './account';
 import { groupSteps } from './groups';
+import { phishingCampaignSteps } from './phishingCampaigns';
 import { trainingCampaignSteps } from './trainingCampaigns';
 import { trainingEnrollmentSteps } from './trainingEnrollments';
 import { userSteps } from './users';
+import { phishingSecurityTestsSteps } from './phishingSecurityTests';
 
 const integrationSteps = [
   ...accountSteps,
@@ -10,6 +12,8 @@ const integrationSteps = [
   ...userSteps,
   ...trainingCampaignSteps,
   ...trainingEnrollmentSteps,
+  ...phishingCampaignSteps,
+  ...phishingSecurityTestsSteps,
 ];
 
 export { integrationSteps };